requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
pygam==0.9.0
matplotlib==3.9.0
PyYAML==6.0.1
tabulate==0.9.0
scipy==1.13.1
mlflow==2.14.1
fastapi==0.111.0
uvicorn==0.30.1
dash==2.17.0
pydantic==2.7.4
requests==2.32.3
# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.7
# Monitoring and observability
psutil==5.9.8
# Testing and quality
pytest==8.0.0
pytest-asyncio==0.23.5
black==24.2.0
flake8==7.0.0
mypy==1.8.0
# Advanced ML
optuna==3.5.0
fairlearn==0.10.0